+++ packages/icon/index.d.ts
@@ -1,9 +1,0 @@
-import { UIElement } from "@elf-framework/sapa";
-
-declare module "@elf-framework/icon" {
-
-}
-
-declare module "@elf-framework/icon/esm/AddChartFilled" {
-    export default class AddChartFilled extends UIElement {}
-}